Rename derive macro
