+++ typesense/tests/derive/ui/duplicated_attribute.rs
@@ -1,6 +1,6 @@
-use typesense::Document;
+use typesense::Typesense;
 use serde::{Serialize, Deserialize};
-#[derive(Document, Serialize, Deserialize)]
+#[derive(Typesense, Serialize, Deserialize)]
 struct Company {
     company_name: String,
     num_employees: i32,
